+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Features
 
 - Add a tutorial for caching quantities and fix a bug in the caching pipeline ([#52](https://github.com/SGIModel/MUSE_OS/pull/52))
+- Update documentation about not using retrofit ([51](https://github.com/SGIModel/MUSE_OS/pull/51))
+- Add error about not finding interaction network ([50](https://github.com/SGIModel/MUSE_OS/pull/50))
 - Add option for `standard_demand` and remove retro agents ([#35](https://github.com/SGIModel/MUSE_OS/pull/35))
 - Update main version ([#28](https://github.com/SGIModel/MUSE_OS/pull/28))
 - Update main version ([#26](https://github.com/SGIModel/MUSE_OS/pull/26))